+++ core/src/main/java/org/realityforge/proton/AnnotationsUtil.java
@@ -8,7 +8,6 @@ import java.util.function.Function;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.lang.model.AnnotatedConstruct;
-import javax.lang.model.SourceVersion;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.AnnotationValue;
 import javax.lang.model.element.Element;
@@ -199,33 +198,12 @@ public final class AnnotationsUtil
   {
     final String declaredName =
       (String) getAnnotationValue( method, annotationClassname, parameterName ).getValue();
-    if ( sentinelValue.equals( declaredName ) )
-    {
-      final String defaultValue = defaultExtractor.apply( method );
-      if ( null == defaultValue )
-      {
-        throw new ProcessorException( MemberChecks.toSimpleName( annotationClassname ) + " target did not specify " +
-                                      "the parameter " + parameterName + " and the default value could not be derived",
-                                      method );
-      }
-      return defaultValue;
-    }
-    else
-    {
-      if ( !SourceVersion.isIdentifier( declaredName ) )
-      {
-        throw new ProcessorException( MemberChecks.toSimpleName( annotationClassname ) + " target specified an " +
-                                      "invalid value '" + declaredName + "' for the parameter " + parameterName + ". " +
-                                      "The value must be a valid java identifier", method );
-      }
-      else if ( SourceVersion.isKeyword( declaredName ) )
-      {
-        throw new ProcessorException( MemberChecks.toSimpleName( annotationClassname ) + " target specified an " +
-                                      "invalid value '" + declaredName + "' for the parameter " + parameterName + ". " +
-                                      "The value must not be a java keyword", method );
-      }
-      return declaredName;
-    }
+    return NamesUtil.extractName( method,
+                                  defaultExtractor,
+                                  annotationClassname,
+                                  parameterName,
+                                  sentinelValue,
+                                  declaredName );
   }
 
   public static boolean hasNonnullAnnotation( @Nonnull final Element element )